add os option to environment create. add linux and windows service ami environment variables to the api

+++ api/backend/ecs/environment_manager.go
@@ -145,9 +145,24 @@ func (this *ECSEnvironmentManager) describeAutoscalingGroup(ecsEnvironmentID id.
 func (this *ECSEnvironmentManager) CreateEnvironment(
 	environmentName string,
 	instanceSize string,
+	operatingSystem string,
 	minClusterCount int,
 	userDataTemplate []byte,
 ) (*models.Environment, error) {
+
+	var defaultUserDataTemplate []byte
+	var serviceAMI string
+	switch strings.ToLower(operatingSystem) {
+	case "linux":
+		defaultUserDataTemplate = defaultLinuxUserDataTemplate
+		serviceAMI = config.AWSLinuxServiceAMI()
+	case "windows":
+		 defaultUserDataTemplate = defaultWindowsUserDataTemplate
+		serviceAMI = config.AWSWindowsServiceAMI()
+	default:
+		return nil, fmt.Errorf("Operating system '%s' is not recognized", operatingSystem)
+	}
+
 	environmentID := id.GenerateHashedEntityID(environmentName)
 	ecsEnvironmentID := id.L0EnvironmentID(environmentID).ECSEnvironmentID()
 
@@ -181,7 +196,6 @@ func (this *ECSEnvironmentManager) CreateEnvironment(
 
 	agentGroupID := config.AWSAgentGroupID()
 	securityGroups := []*string{groupID, &agentGroupID}
-	serviceAMI := config.AWSServiceAMI()
 	ecsRole := config.AWSECSInstanceProfile()
 	keyPair := config.AWSKeyPair()
 	launchConfigurationName := ecsEnvironmentID.LaunchConfigurationName()
@@ -374,8 +388,82 @@ func renderUserData(ecsEnvironmentID id.ECSEnvironmentID, userData []byte) (stri
 	return base64.StdEncoding.EncodeToString(rendered.Bytes()), nil
 }
 
-var defaultUserDataTemplate = []byte(
-	`#!/bin/bash
+// todo: this script is still WIP
+
+var defaultWindowsUserDataTemplate = []byte(
+`<powershell>
+# Set agent env variables for the Machine context (durable)
+$clusterName = "{{ .ECSEnvironmentID }}"
+Write-Host Cluster name set as: $clusterName -foreground green
+
+[Environment]::SetEnvironmentVariable("ECS_CLUSTER", $clusterName, "Machine")
+[Environment]::SetEnvironmentVariable("ECS_ENABLE_TASK_IAM_ROLE", "false", "Machine")
+$agentVersion = 'v1.14.0-1.windows.1'
+$agentZipUri = "https://s3.amazonaws.com/amazon-ecs-agent/ecs-agent-windows-$agentVersion.zip"
+$agentZipMD5Uri = "$agentZipUri.md5"
+
+# Configure docker auth
+Read-S3Object -Region {{ .Region }} -BucketName {{ .S3Bucket }} -Key bootstrap/dockercfg -File dockercfg
+
+
+$dockercfgContent = '{"d.ims.io":{"auth":"<***PRIVATE_REG_AUTH_TOKEN***>","email":""}}';
+[Environment]::SetEnvironmentVariable("ECS_ENGINE_AUTH_DATA", $dockercfgContent, "Machine")
+[Environment]::SetEnvironmentVariable("ECS_ENGINE_AUTH_TYPE", "dockercfg", "Machine")
+
+# Output all environment variables
+Get-ChildItem -Path Env:* | Sort-Object Name
+
+# Output all environment variables
+Get-ChildItem -Path Env:* | Sort-Object Name
+
+### --- Nothing user configurable after this point ---
+$ecsExeDir = "$env:ProgramFiles\Amazon\ECS"
+$zipFile = "$env:TEMP\ecs-agent.zip"
+$md5File = "$env:TEMP\ecs-agent.zip.md5"
+
+### Get the files from S3
+Invoke-RestMethod -OutFile $zipFile -Uri $agentZipUri
+Invoke-RestMethod -OutFile $md5File -Uri $agentZipMD5Uri
+
+## MD5 Checksum
+$expectedMD5 = (Get-Content $md5File)
+$md5 = New-Object -TypeName System.Security.Cryptography.MD5CryptoServiceProvider
+$actualMD5 = [System.BitConverter]::ToString($md5.ComputeHash([System.IO.File]::ReadAllBytes($zipFile))).replace('-', '')
+
+if($expectedMD5 -ne $actualMD5) {
+    echo "Download doesn't match hash."
+    echo "Expected: $expectedMD5 - Got: $actualMD5"
+    exit 1
+}
+
+## Put the executables in the executable directory.
+Expand-Archive -Path $zipFile -DestinationPath $ecsExeDir -Force
+
+## Start the agent script in the background.
+$jobname = "ECS-Agent-Init"
+$script =  "cd '$ecsExeDir'; .\amazon-ecs-agent.ps1"
+$repeat = (New-TimeSpan -Minutes 1)
+
+$jobpath = $env:LOCALAPPDATA + "\Microsoft\Windows\PowerShell\ScheduledJobs\$jobname\ScheduledJobDefinition.xml"
+if($(Test-Path -Path $jobpath)) {
+  echo "Job definition already present"
+  exit 0
+
+}
+
+$scriptblock = [scriptblock]::Create("$script")
+$trigger = New-JobTrigger -At (Get-Date).Date -RepeatIndefinitely -RepetitionInterval $repeat -Once
+$options = New-ScheduledJobOption -RunElevated -ContinueIfGoingOnBattery -StartIfOnBattery
+Register-ScheduledJob -Name $jobname -ScriptBlock $scriptblock -Trigger $trigger -ScheduledJobOption $options -RunNow
+Add-JobTrigger -Name $jobname -Trigger (New-JobTrigger -AtStartup -RandomDelay 00:1:00)
+</powershell>
+<persist>true</persist>
+`)
+
+
+
+var defaultLinuxUserDataTemplate = []byte(
+`#!/bin/bash
     echo ECS_CLUSTER={{ .ECSEnvironmentID }} >> /etc/ecs/ecs.config
     echo ECS_ENGINE_AUTH_TYPE=dockercfg >> /etc/ecs/ecs.config
     yum install -y aws-cli awslogs jq
